backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "RAG system for semantic search of real estate properties in Mexico"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.133.1",
    "google-genai>=1.65.0",
    "httpx>=0.28.1",
    "openai>=2.24.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.17.0",
    "uvicorn[standard]>=0.41.0",
]
